Don't close reg window before setting webview src
In chrome os 71 there is a slight difference in ordering of async
operations.

The createWebviewWindow promise resolves before the viewer window
DOMContentLoaded event. This results in the registration window being
closed due to the previousWindow.close() call and so the webview.src is
never set and viewer.risevision.com/Viewer.html is never loaded.

That in turn results in the script injection not firing because it is
waiting for loadCommit event.